+++ CHANGELOG.md
@@ -14,6 +14,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add `Debug` implementation for `ua::Array<T>` data types.
 - Add `ValueType` enum to check `ua::Variant` without unwrapping (also `ua::Argument`).
 - Add tracing log messages when processing service requests and responses.
+- Add methods to `ClientBuilder` to set response timeout, client description, and connectivity check
+  interval.
 
 ### Changed
 
@@ -23,6 +25,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Breaking: Move `ua::VariantValue` and `ua::ScalarValue` to top-level export outside `ua`
 - Breaking: Remove `ua::ArrayValue` for now (until we have a better interface).
 - Breaking: Return output arguments directly from `AsyncClient::call_method()`, without `Option`.
+- Breaking: Remove misleading `FromStr` trait implementation and offer `ua::String::new()` instead.
 
 ### Fixed
 
@@ -33,6 +36,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   to `ua::StatusCode::GOOD`.
 - No longer panic when unwrapping `ua::Variant` with array value.
 - Allow invalid references array in `ua::BrowseResult` when request was otherwise successful.
+- Handle graceful disconnection when dropping synchronous `Client`.
 
 ## [0.5.0] - 2024-03-01